pyproject.toml
[project]
name = "firstcall"
version = "0.1.0"
description = "AI voice agent for medical emergency triage and first aid guidance"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "openai>=1.54.0",
    "twilio>=9.3.5",
    "deepgram-sdk>=3.7.0",
    "elevenlabs>=1.9.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "httpx>=0.27.2",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
